* main: (110 commits)
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release inception-34.0
  No issue: Mention MS SQL and PosgreSQL as experimental DB options
  No issue: Fix versions after merge
  #5043 - Ability to specify token breaking zones when calling tokenizer
  #5064 - Project template for PHI annotation
  #5071 - Better document which layers are supported by which formats
  #5061 - Multiple synchronous recommenders only the last one wins
  #5064 - Project template for PHI annotation
  #5068 - Show annotation sidebar by default
  #5033 - Ability to configure recommenders interactively on the annotation page
  #5066 - Increase default number of rows for brat editors
  #5056 - Ability to configure additional languages for knowledge bases
  #4909 - Upgrade dependencies
  #5056 - Ability to configure additional languages for knowledge bases
  #4909 - Upgrade dependencies
  No issue: Actually, document-metadata doesn't seem to be experimental after all...
  No issue: Set version to 35.0-SNAPSHOT
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release inception-33.7
  ...

+++ inception/inception-documents/src/main/java/de/tudarmstadt/ukp/inception/documents/DocumentStorageServiceImpl.java
@@ -81,7 +81,9 @@ public class DocumentStorageServiceImpl
         Validate.notNull(aDocument, "Parameter [sourceDocument] must be specified");
 
         var path = getSourceDocumentFolder(aDocument);
-        FileUtils.forceDelete(path.toFile());
+        if (Files.exists(path)) {
+            FileUtils.forceDelete(path.toFile());
+        }
     }
 
     @Override
